Finished bugprone and llvm

+++ libff/algebra/curves/mnt/mnt6/mnt6_pairing.cpp
@@ -265,7 +265,7 @@ mnt6_affine_ate_G2_precomputation mnt6_affine_ate_precompute_G2(const mnt6_G2& Q
     bool found_nonzero = false;
 
     std::vector<long> NAF = find_wnaf(1, loop_count);
-    for (unsigned long i = NAF.size() - 1; i >= 0; --i)
+    for (long i = (long) NAF.size() - 1; i >= 0; --i)
     {
         if (!found_nonzero)
         {
@@ -339,7 +339,7 @@ mnt6_Fq6 mnt6_affine_ate_miller_loop(const mnt6_affine_ate_G1_precomputation &pr
     size_t idx = 0;
 
     std::vector<long> NAF = find_wnaf(1, loop_count);
-    for (unsigned long i = NAF.size() - 1; i >= 0; --i)
+    for (long i = (long) NAF.size() - 1; i >= 0; --i)
     {
         if (!found_nonzero)
         {
@@ -512,7 +512,7 @@ mnt6_ate_G2_precomp mnt6_ate_precompute_G2(const mnt6_G2& Q)
 
     const bigint<mnt6_Fr::num_limbs> &loop_count = mnt6_ate_loop_count;
     bool found_one = false;
-    for (unsigned long i = loop_count.max_bits() - 1; i >= 0; --i)
+    for (long i = (long) loop_count.max_bits() - 1; i >= 0; --i)
     {
         const bool bit = loop_count.test_bit(i);
 
@@ -567,7 +567,7 @@ mnt6_Fq6 mnt6_ate_miller_loop(const mnt6_ate_G1_precomp &prec_P,
 
     const bigint<mnt6_Fr::num_limbs> &loop_count = mnt6_ate_loop_count;
 
-    for (unsigned long i = loop_count.max_bits() - 1; i >= 0; --i)
+    for (long i = (long) loop_count.max_bits() - 1; i >= 0; --i)
     {
         const bool bit = loop_count.test_bit(i);
 
@@ -628,7 +628,7 @@ mnt6_Fq6 mnt6_ate_double_miller_loop(const mnt6_ate_G1_precomp &prec_P1,
 
     const bigint<mnt6_Fr::num_limbs> &loop_count = mnt6_ate_loop_count;
 
-    for (unsigned long i = loop_count.max_bits() - 1; i >= 0; --i)
+    for (long i = (long) loop_count.max_bits() - 1; i >= 0; --i)
     {
         const bool bit = loop_count.test_bit(i);
